Added unit test fir dataset object, fiddled with the linkages, so now:
d = dataset(...)
d.getLatice_info()

d.setLattice('oP')
d.getLattice_info()

will automatically regenerate all of the information. Idea is that when it's
all in place

d.getScaling_info()

will be able to wind back to the start... :o)

Also added reset() method to the core Object to implement this.