(biggie)
Now make it so that the data used for running pointless is only the
first 180 degrees of any sweep - this will prevent the huge reflection
files from e.g. ed mitchell data from clogging up xia2 for ages.

For the spacegroup determination, if there is only one epoch then this
is also done, else what happens is the whole data set is used - you can't
run rebatch on reflection files with many sweeps in so there's no point
trying.

If the null integrater is being used then none of this is done because no
phi information is available - this seems to cope ok.