Changes towards getting XDS supported in xia2 - will need to include CORRECT in the integrater because
(1) it will feed back teh refined orientation
(2) it will feed back the resolution limit

which means, as it assumes that the input pointgroup is correct, that I will
have to check that the pointgroup doesn't make too much of a difference
to the results.

wrappers changed by including the INP files.