Implemented machinery to get the REFERENCE_REFLECTION_FILE from a CRYSTAL to the Scaler - this now needs to be implemented in the Scaler implementations but is a fair amount of the way there...
This should be an MTZ file, and if it has a FreeR column this will be copied
to the output reflection file. This will also be used as a reference for
determining the correct indexing and will also be used to give the correct
spacegroup.

This kind of thing should be added into AMI. Perhaps now is the moment to be
using AMI for the analysis in the two scaler implementations.