Big commit.
Adding support for the reference reflection file.. this should work
for XDS and mosflm/scala. Note well that this is what it does:

use it as an indexing reference
copy the spacegroup from it

I still need to add copying the FreeR column.

Assumes

same lattice
cell is isomorphous (tested to 10%)
reference file contains I columns