Sundry changes to get the xia2 working with small molecule data (the RJ commits got caught up in the nets) -
if small_molecule set n_bins to a factor of 10 smaller number for resolution
estimation - N.B. this is limited to 0.707A at the moment, should fix. Also
don't run truncate on this, as that will probably be meaningless. Also am
not using a Wilson method to do any fits...