A fix I think for the trac ticket #168:
P.S. It should also probably be clarified how -freer_files interact with -spacegroup and the normal indexing, as keeping a -freer_file implicitly forces the space group to match the previous mtz file. Currently, if you have a -freer_file specified the indexing proceeds as normal (not necessarily converging on the desired space group) and the data gets reindexed(?) after scaling. I would imagine that, if a -freer_file is specified, you should extract the space group of that mtz file and then impose that space group *as if the user had specified it using -spacegroup*. Does that make sense?