Change semantics for aimless scale factor output file.
The aimless wrapper will request aimless to write scale factors to the specified file.
Do not raise exceptions if the file does not exist after aimless was run.
Users of the wrapper should use get_new_scales_file() to retrieve the file name, which will
then run a check and raise an appropriate exception if the file is not present.