Compare start and end image bounds with range of found images
Ensure that the user-specified start and end images lie fully within
the range of found images in a given sweep, before populating the
.xinfo file.  Reject a sweep if it doesn't meet such a condition, and
print an explanatory note in the debug log.

For a sweep with missing images, which xia2 identifies as multiple
sweeps, this prevents duplicate entries in the .xinfo file.
Fixes #240.