Update expected result for x4-wide dials_full
Update the I/sigma expected results. Although there is some slight platform dependence likely to originate in dials.refine, I think the fact that this leads to a large variation in I/sigma is because this is quite a 'bad' dataset, so slight variations in the error model fitting in scaling lead to quite large changes in average I/sigma. Therefore I am setting the expected values to the average of the different results with a high uncertainty, to reflect the fact that is is likely that the I/sigma truly does have a high dependency on the exact processing steps due to the dataset quality.