Update command line tool log streaming interface
Update all command line tools that import xia2 components to explicitly
create named log files and redirect all Chatter() and other stream
output to said log files.

Some commands that do not import from xia2 were not touched.

This is the first half of the first step of the list in #267