Reset the all integraters of a reset refiner
In the case of multi-sweep refinement, a single refiner can have
multiple associated integraters.  If the refiner has been reset, we
need to reset all the associated integraters at once.  If we don't, the
first integrater will be reset, but the second and all subsequent
integraters' calls to get_integrater_prepare_done will find
get_refiner_done is true and they won't be reset.