Prevent unit cell restraints from being disabled
Scan-varying refinement causes restraints to be disabled unless we
specify a scan-static unit cell in the call to dials.refine.  We should
always do this if the xia2.Wrappers.Dials.Refine.Refine instance's
attribute _scan_varying is True and one or more of its attributes
tie_to_group or tie_to_target are truthy.