Use better refinement params for narrow wedges
The default width of intervals between checkpoints for scan-varying
refinement is 36°.  For scans narrower than this, we aim in xia2 to
ensure a minimum of two intervals (three checkpoints) by default.
Previously this was achieved by querying the total scan angle and
halving it.  However, dials.refine may trim the image range internally
if there are no reflections at the start or end of a scan, which may
leave one or both checkpoint intervals partly or completely empty.  To
avoid resultant scan-varying refinement errors, we can instead use the
parameter absolute_num_intervals=2 in dials.refine, which is more robust
to the scan being trimmed.