One interval makes more sense that two
For consistency with behaviour at a scan width of 36°, where the default
checkpoint interval applies and leads to scan-varying refinement using
two checkpoints (one interval), it doesn't make sense for a scan width
less than 36° to result in a step change to two intervals.  Instead, we
should fall back on a single interval for narrow wedges.