Fix load_imagesets introduced by #491 (#499)
* Don't overwrite input kwarg

This mean that load_imageset only returned one of the imagesets in
cases that it should have returned multiple imagesets with the same
template.

Fixes #498.

* Bug fix for sorting sweeps with matching epochs

I.e. two sweeps with the same template will have the same epoch,
resulting in the following error:

TypeError: '<' not supported between instances of 'Sweep' and 'Sweep'

* Reinstate xia2.setup command

Reads command line, generates xinfo file, and that's it.

* Use xia2.setup in test_xia2setup

Then we can call via procrunner and avoid having to workaround
various xia2 singleton instances.