Remove tests from the xia2 python package
Tests are moved out of src/ into a top-level tests directory.
This means we no longer distribute tests as part of the release package.
Which is fine, since the release package already doesn't contain the
pytest configuration file setting up test fixtures.

If someone wants to run tests they can still do so by running tests on
the packed up sources relating to a release
(but not the sdist/wheel/installed package).

Update the package definition to partially use a configuration file.