Allow installation of xia2 with read-only conda_base/ (#616)
On some HPC systems, the root environment is read-only. This allows
installation and usage within the libtbx environment without having to
create extra layers of virtual environments.

It does this by falling back to inserting the editable source PYTHONPATH
into the tbx dispatcher system's list of extra python paths.

Notably, xia2 doesn't have the vestigial equivalent of the
dxtbx/src/dxtbx/command_line directory, so there is no place to trigger
dispatcher generation. This means that this installed xia2 won't have
the `xia2` dispatcher generated. Given the use-case of the target
systems for this, I think it more that installation fails that is the
problem, so this might not be an issue.