Use smaller datasets in some regression tests
Instead of the relatively large 'insulin' dataset use
'centroid-test-data' for some regression tests.

This reduces test times between 50% and 75% for the modified tests,
and allows more tests to be enabled to run without --regression-full.

Two tests were enabled to run on Azure (one due to runtime reduction,
the other due to runtime reduction and the removal of an incorrect CCP4
dependency declaration)