xia2.multiplex: Handle multiple wavelengths (#755)
This allows xia2.multiplex to handle multi-wavelength data.

The default wavelength tolerance can be increased; if all data is within
the tolerance it will be processed and saved together, writing the
average wavelength to the output MTZ files.

If there are wavelength differences greater than the tolerance, the
multiplex analysis is still performed on all data as usual, however upon
saving the data the MTZ, separate files are written e.g. scaled_WAVE1.mtz.
Statistics reporting/html output is still done on the data as a whole.