Fix incorrect handling of Windows drive names.
The previous method only worked if the user was trying to do an import
with a :start:end slice at the end. If there was no slice, only a drive
letter, then the old method failed with

Error: /path/to/image.h5:start:end

With this change, extract the drive letter part of the path first
before splitting into tokens.